Renamed typename to toTypeName and added documentation for it
