+++ StepLang/StepLang.csproj
@@ -29,4 +29,8 @@
         <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       </None>
     </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Framework\Reflection\" />
+    </ItemGroup>
 </Project>
